requirements.txt
pycord
python-dotenv
requests
